requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-core
langchain-chroma
python-dotenv
zhipuai
unstructured
pymupdf